Refactored out common dependencies
