+++ ci.Dockerfile
@@ -8,11 +8,15 @@ RUN chmod +x /usr/bin/dep
 RUN go get -u github.com/alecthomas/gometalinter
 RUN gometalinter --install
 
-#Restore dep for dispatcher
+# Copy common dependencies
+WORKDIR /go/src/github.com/lawrencegripper/ion/common
+COPY common .
+
+# Restore dep for dispatcher
 WORKDIR /go/src/github.com/lawrencegripper/ion/$FOLDER
-COPY ./$FOLDER/Gopkg.lock .
-COPY ./$FOLDER/Gopkg.toml .
+COPY ./${FOLDER}/Gopkg.lock .
+COPY ./${FOLDER}/Gopkg.toml .
 RUN dep ensure -vendor-only
-COPY ./$FOLDER .
+COPY ${FOLDER} .
 RUN go test -v -race -short ./...
 RUN gometalinter --vendor --disable-all --enable=errcheck --enable=vet --enable=gofmt --enable=golint --enable=deadcode --enable=varcheck --enable=structcheck --deadline=15m ./...
