+++ sidecar/types/types.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"net/http"
 
-	"github.com/lawrencegripper/ion/dispatcher/messaging"
+	"github.com/lawrencegripper/ion/common"
 )
 
 //MetadataProvider is a document storage DB for storing document data
@@ -24,26 +24,26 @@ type BlobProvider interface {
 
 //EventPublisher is responsible for publishing events to a remote system
 type EventPublisher interface {
-	Publish(e messaging.Event) error
+	Publish(e common.Event) error
 	Close()
 }
 
 //EventContext is a single entry in a document
 type EventContext struct {
-	EventID       string                   `bson:"id" json:"id"`
-	CorrelationID string                   `bson:"correlationId" json:"correlationId"`
-	ParentEventID string                   `bson:"parentEventId" json:"parentEventId"`
-	Files         []string                 `bson:"files" json:"files"`
-	Data          []messaging.KeyValuePair `bson:"data" json:"data"`
+	EventID       string                `bson:"id" json:"id"`
+	CorrelationID string                `bson:"correlationId" json:"correlationId"`
+	ParentEventID string                `bson:"parentEventId" json:"parentEventId"`
+	Files         []string              `bson:"files" json:"files"`
+	Data          []common.KeyValuePair `bson:"data" json:"data"`
 }
 
 //Insight todo
 type Insight struct {
-	ExecutionID   string                   `bson:"id" json:"id"`
-	CorrelationID string                   `bson:"correlationId" json:"correlationId"`
-	EventID       string                   `bson:"eventId" json:"eventId"`
-	ParentEventID string                   `bson:"parentEventId" json:"parentEventId"`
-	Data          []messaging.KeyValuePair `bson:"data" json:"data"`
+	ExecutionID   string                `bson:"id" json:"id"`
+	CorrelationID string                `bson:"correlationId" json:"correlationId"`
+	EventID       string                `bson:"eventId" json:"eventId"`
+	ParentEventID string                `bson:"parentEventId" json:"parentEventId"`
+	Data          []common.KeyValuePair `bson:"data" json:"data"`
 }
 
 //ErrorResponse is a struct intended as JSON HTTP response
